Move location of roundrobin to be under /peerlist/ & small fixes
